DMN retains the old feel-like behavior
All other scripts are now strictly python only.
