+++ tests/SpiffWorkflow/bpmn/TimerDurationBoundaryTest.py
@@ -8,7 +8,7 @@ import time
 from SpiffWorkflow.task import Task
 from SpiffWorkflow.bpmn.workflow import BpmnWorkflow
 from tests.SpiffWorkflow.bpmn.BpmnWorkflowTestCase import BpmnWorkflowTestCase
-
+from SpiffWorkflow.bpmn.BpmnFeelScriptEngine import BpmnFeelScriptEngine
 __author__ = 'kellym'
 
 
@@ -29,6 +29,7 @@ class TimerDurationTest(BpmnWorkflowTestCase):
 
     def actual_test(self,save_restore = False):
         self.workflow = BpmnWorkflow(self.spec)
+        self.workflow.script_engine = BpmnFeelScriptEngine()
         ready_tasks = self.workflow.get_tasks(Task.READY)
         self.assertEqual(1, len(ready_tasks))
         self.workflow.complete_task_from_id(ready_tasks[0].id)
@@ -47,7 +48,9 @@ class TimerDurationTest(BpmnWorkflowTestCase):
             ready_tasks = self.workflow.get_tasks(Task.READY)
             if len(ready_tasks) < 1:
                 break
-            if save_restore: self.save_restore()
+            if save_restore:
+                self.save_restore()
+                self.workflow.script_engine = BpmnFeelScriptEngine()
             #self.assertEqual(1, len(self.workflow.get_tasks(Task.WAITING)))
             time.sleep(0.1)
             self.workflow.complete_task_from_id(ready_tasks[0].id)
